Move unremovable to AsclepiusSystem.Staff
